Task_With_Docker/requirements.txt
mysql-connector-python==8.0.20
requests==2.22.0
flask==1.1.2